chore: eventCss
