+++ src/app/components/dashboards/communication/view/witness-form/witness-form.component.html
@@ -1,4 +1,4 @@
-<form [formGroup]="witnessForm" (ngSubmit)="onSubmitWitnessForm()">
+ <form [formGroup]="witnessForm" (ngSubmit)="onSubmitWitnessForm()">
   <div class="example-container">
     <div class="imgContainer">
       <img src="../../../../../../assets/images/close.png" (click)="onClose()" />
@@ -18,7 +18,7 @@
 
     <div class="container">
       <button mat-flat-button type="button" (click)="resetForm()">Effacer</button>
-      <button mat-flat-button type="submit">Envoyer</button>
+      <button mat-flat-button type="submit" [disabled]="!witnessForm.valid">Envoyer</button>
     </div>
   </div>
 </form>
