--- src/app/components/dashboards/evenements/evenements.component.scss
@@ -1,29 +1,0 @@
-body{
-  height: 100%;
-}
-.box-essai{
-  height: 100%;
-  width: 100%;
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-evenly;
-
-}
-
-.essai{
-  color: black;
-  height: auto;
-  width: 40%;
-  background-color: #ffff8a;
-  margin: 15px;
-  display: flex;
-  text-align: center;
-  align-items: center;
-  justify-content: center;
-  padding: 5px;
-
-  img{
-    width: 100%;
-    height: 100%;
-  }
-}
