--- src/app/components/dashboards/evenements/evenements.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-evenements',
-  templateUrl: './evenements.component.html',
-  styleUrls: ['./evenements.component.scss']
-})
-export class EvenementsComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
