--- src/app/components/dashboards/events/events.component.html
@@ -1,14 +1,0 @@
-<mat-tab-group>
-  <mat-tab label="Témoignages">
-    <div class="section mat-elevation-z4" *ngFor="let item of witness">
-      <div class="texte" >
-        <h4>{{item.username}} -- {{item.note}}/10 </h4>
-        <h3> {{item.title}}</h3>
-        <p>{{item.description}}</p>
-      </div>
-      <div>
-        <img src="http://via.placeholder.com/60x60" alt="placeholder">
-      </div>
-     </div>
-  </mat-tab>
-</mat-tab-group>
