+++ src/app/pages/dashboard/communication/communication.component.scss
@@ -5,13 +5,13 @@
   margin-top: 5vh;
 }
 .leftSide{
-  width: 40%;
+  width: 50%;
   display: flex;
   flex-direction: column;
 }
 .rightSide{
   background-color: white;
-  width: 40%;
+  width: 50%;
 }
 .invisible{
   display: none;
@@ -27,4 +27,13 @@
     justify-content: center;
   }
 
+}
+.leftSide{
+  width: 40%;
+  display: flex;
+  flex-direction: column;
+}
+.rightSide{
+  background-color: white;
+  width: 40%;
 }
