+++ src/app/pages/dashboard/events/events.component.html
@@ -1,0 +1,14 @@
+<div class="container">
+  <mat-tab-group>
+    <mat-tab label="Evenements à venir">
+      <div class="section mat-elevation-z4" *ngFor="let item of events">
+        <div class="texte" >
+          <h4>Le {{item.date |date :'dd/MM/yyyy'}} à {{item.date | date : 'HH:mm'}}</h4>
+          <h3> {{item.title}}</h3>
+          <p>{{item.description}}</p>
+          <small>--{{item.author}}--</small>
+        </div>
+       </div>
+    </mat-tab>
+  </mat-tab-group>
+</div>
