+++ src/app/pages/dashboard/events/events.component.scss
@@ -1,0 +1,21 @@
+.container{
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  margin-top: 5vh;
+  width: 80%;
+  background-color: white;
+}
+
+.section{
+  display: flex;
+  background-color: white;
+  padding: 16px;
+  border-radius: 3px;
+  margin: 20px;
+}
+mat-tab-group{
+  max-height: 80vh;
+  width: 100%;
+  border-radius: 3px;
+}
